# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Transaction": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string",
						"enum": [
							"credit",
							"debit"
						]
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"type",
					"amount",
					"timestamp",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"balance": {
						"type": "number",
						"format": "double"
					},
					"googleId": {
						"type": "string"
					},
					"transactions": {
						"items": {
							"$ref": "#/components/schemas/Transaction"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"balance",
					"googleId",
					"transactions"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "fintech-wallet",
		"version": "1.0.0",
		"description": "A fintech wallet API built with Node.js, Express, and PostgreSQL.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/paypal/create": {
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"200": {
						"description": "Order created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"amount": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"amount"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/paypal/capture": {
			"post": {
				"operationId": "CaptureOrder",
				"responses": {
					"200": {
						"description": "Order captured",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"newBalance": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"newBalance",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"userId": {
										"type": "number",
										"format": "double"
									},
									"orderId": {
										"type": "string"
									}
								},
								"required": [
									"userId",
									"orderId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/paypal/transfer": {
			"post": {
				"operationId": "TransferFunds",
				"responses": {
					"200": {
						"description": "Transfer successful",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"recipientBalance": {
											"type": "number",
											"format": "double"
										},
										"senderBalance": {
											"type": "number",
											"format": "double"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"recipientBalance",
										"senderBalance",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"amount": {
										"type": "number",
										"format": "double"
									},
									"recipientEmail": {
										"type": "string"
									},
									"senderId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"amount",
									"recipientEmail",
									"senderId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/paypal/transactions": {
			"get": {
				"operationId": "GetTransactions",
				"responses": {
					"200": {
						"description": "Transactions fetched",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Transaction"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}